Setup admin_controller, admin folders, refactor user & application controller, routes, I18n and views
