+++ app/views/admin/users/index.html.erb
@@ -3,6 +3,5 @@
   <% @users.each do |user| %>
     <li><%= user.name %></li>
     <li><%= user.email %></li>
-    </br>
   <% end %>
 </ul>
