+++ app/views/admin/users/new.html.erb
@@ -1,5 +1,5 @@
 <h1><%= t(".title") %></h1>
-<%= simple_form_for @user do |f| %>
+<%= simple_form_for [:admin, @user] do |f| %>
   <%= f.input :name %>
   <%= f.input :email %>
   <%= f.input :password %>
