+++ app/views/home/show.html.erb
@@ -2,4 +2,4 @@
 
 <h2><%= t(".new_account") %></h2>
 
-<%= link_to t(".link"), new_user_path %>
+<%= link_to t(".link"), new_admin_user_path %>
