+++ config/locales/en.yml
@@ -69,22 +69,23 @@ en:
       flash:
         notice: "Logout successful"
 
-  users:
-    defaults:
-      name: "Name"
-      email: "Email"
-      password: "Password"
-    index:
-      require_admin: "You must be admin to access user list"
-      title: "Current registered users"
-    new:
-      title: "Please fill in the form to sign up"
-    edit:
-      delete: "Delete Account"
-      title: "Please fill in the form to edit your profile"
-    show:
-      profile_information: "Your email address is: %{email}"
-      welcome: "Welcome, your details are:"
+  admin:      
+    users:
+      defaults:
+        name: "Name"
+        email: "Email"
+        password: "Password"
+      index:
+        require_admin: "You must be admin to access user list"
+        title: "Current registered users"
+      new:
+        title: "Please fill in the form to sign up"
+      edit:
+        delete: "Delete Account"
+        title: "Please fill in the form to edit your profile"
+      show:
+        profile_information: "Your email address is: %{email}"
+        welcome: "Welcome, your details are:"
 
   activerecord:
     errors:
