+++ config/routes.rb
@@ -1,6 +1,9 @@
 Rails.application.routes.draw do
   resources :budgets
   resources :sessions, only: [:new, :create, :destroy]
-  resources :users
+  namespace :admin do
+    resources :users
+  end
+  # resources :users, only: [:new, :create, :show, :edit, :update]
   root to: "home#show"
 end
